pyproject.toml
[project]
name = "lithium"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django~=6.0",
    "django-allauth[openid,socialaccount]~=65.13",
    "django-debug-toolbar~=6.1",
    "crispy-bootstrap5~=2025.6",
    "gunicorn~=23.0",
    "psycopg[binary]~=3.3",
    "whitenoise~=6.11",
    "django-environ~=0.13.0",
]

[dependency-groups]
dev = [
    "typing_extensions~=4.15",
    "ipython~=9.8",
    "ruff~=0.14",
    "django-dev-superuser~=0.2.0",
]

# ------------------------------------------ #
# ------------------ Ruff ------------------ #
# ------------------------------------------ #
[tool.ruff]
line-length = 119
exclude = ["migrations"]
fix = true

[tool.ruff.lint]
fixable = ["ALL"]
select = [
    "C4", # flake8-comprehensions
    "E", # pycodestyle - Errors
    "F", # Pyflakes
    "I", # isort
    "SIM", # flake8-simplify (SIM)
    "T20", # flake8-print
    "UP", # pyupgrade
    "W", # pycodestyle - Warnings
]

ignore = []
[tool.ruff.lint.per-file-ignores]
# https://docs.astral.sh/ruff/settings/#per-file-ignores
"config/settings/*" = ["F405"]

[tool.ruff.lint.isort]
force-wrap-aliases = true
combine-as-imports = true
case-sensitive = true
relative-imports-order = "closest-to-furthest"
